More code cleanup and refactoring
Added NOOP definition.
Cleaned up the logging calls.
Decided to remove the extra type defines and replace them with the std _t's.
Default tilemap scaling is now based of virtual coordinates.
Misc Clang warnings handled.
Renamed "auto" tilemap scale property to "screen".
Updated a few demo to new default virtual size.
Updated TinyXML2.

+++ src/graphics/dummy/texturecoordinatedata.cpp
@@ -42,8 +42,8 @@ using namespace Dummy;
 
 const Core::Type TextureCoordinateData::sType("Graphics::TextureCoordinateData");
 
-TextureCoordinateData::TextureCoordinateData(UINT16 c)
-#define AXES 2u
+TextureCoordinateData::TextureCoordinateData(uint16_t c)
+#define AXES 2
     : m_data(new float[c * AXES]) // TODO: replace with custom allocator
     , m_count(c)
 {
@@ -56,18 +56,18 @@ TextureCoordinateData::~TextureCoordinateData(void)
 }
 
 bool
-TextureCoordinateData::get(UINT16 i, float &u, float &v) const
+TextureCoordinateData::get(uint16_t i, float &u, float &v) const
 {
-	const UINT16 l_offset = (i % m_count) * AXES;
+	const uint16_t l_offset = static_cast<uint16_t>((i % m_count) * AXES);
 	u = m_data[l_offset];
 	v = m_data[l_offset + 1];
 	return(true);
 }
 
 bool
-TextureCoordinateData::set(UINT16 i, float u, float v)
+TextureCoordinateData::set(uint16_t i, float u, float v)
 {
-	const UINT16 l_offset = (i % m_count) * AXES;
+	const uint16_t l_offset = static_cast<uint16_t>((i % m_count) * AXES);
 	m_data[l_offset] = u;
 	m_data[l_offset + 1] = v;
 	return(true);
