+++ src/graphics/opengl/vertexdata.cpp
@@ -44,8 +44,8 @@ using namespace OpenGL;
 
 const Core::Type VertexData::sType("Graphics::VertexData");
 
-VertexData::VertexData(UINT16 c)
-#define AXES 2u
+VertexData::VertexData(uint16_t c)
+#define AXES 2
     : m_id()
     , m_data(new GLfloat[c * AXES]) // TODO: replace with custom allocator
     , m_count(c)
@@ -90,18 +90,18 @@ VertexData::unbuffer(void)
 }
 
 bool
-VertexData::get(UINT16 i, float &x, float &y) const
+VertexData::get(uint16_t i, float &x, float &y) const
 {
-	const int l_offset = (i % m_count) * AXES;
+	const uint16_t l_offset = static_cast<uint16_t>((i % m_count) * AXES);
 	x = m_data[l_offset];
 	y = m_data[l_offset + 1];
 	return(true);
 }
 
 bool
-VertexData::set(UINT16 i, float x, float y)
+VertexData::set(uint16_t i, float x, float y)
 {
-	const UINT16 l_offset = (i % m_count) * AXES;
+	const uint16_t l_offset = static_cast<uint16_t>((i % m_count) * AXES);
 	m_data[l_offset] = x;
 	m_data[l_offset + 1] = y;
